refactor: updated validation schemas to .json over .py
