+++ grading-vm/orca_grader/common/grading_job/grading_job_output.py
@@ -24,7 +24,7 @@ class GradingJobOutput:
     json_responses = list(map(lambda c: c.to_json(), self.__command_responses))
     ans["shell_responses"] = json_responses
     if self.__execution_errors is not None:
-      ans["execution_errors"] = list(map(lambda e: str(e), self.__execution_errors))
+      ans["execution_errors"] = list(map(lambda e: f"{e.__class__.__name__}: {e}", self.__execution_errors))
     if self.__tap_output is not None:
       ans["tap_output"] = self.__tap_output
     return ans
