+++ grading-vm/orca_grader/tests/fixtures/files/live-URL-student-only.json
@@ -1,5 +1,5 @@
 {
-  "target": {
+  "target_code": {
     "url": "https://handins.ccs.neu.edu/files/36/831/FM/FMACgaXsE2coOqvffGX9FA/submission/HW04.zip",
     "mime_type": "application/zip"
   },
