+++ grading-vm/orca_grader/validations/schemas/bash_grading_script_command_schema.json
@@ -1,4 +1,4 @@
-BashGradingScriptCommandSchema = {
+{
   "type": "object",
   "properties": {
     "cmd": { "type": "string" },
