+++ grading-vm/orca_grader/validations/schemas/code_file_schema.json
@@ -1,4 +1,4 @@
-CodeFileSchema = {
+{
   "type": "object",
   "properties": {
     "url": { "type": "string" },
