+++ grading-vm/orca_grader/validations/schemas/conditional_grading_script_comamnd_schema.json
@@ -1,8 +1,8 @@
-ConditionalGradingScriptCommandSchema = {
+{
   "type": "object",
   "properties": {
-    "condition": { 
-      "$ref": "validations/schemas/grading_script_condition_schema.py" 
+    "condition": {
+      "$ref": "grading_script_condition_schema.json"
     },
     "on_true": { "type": ["string", "number"] },
     "on_false": { "type": ["string", "number"] }
