+++ grading-vm/orca_grader/validations/schemas/grading_script_condition_schema.json
@@ -1,4 +1,4 @@
-GradingScriptConditionSchema = {
+{
   "type": "object",
   "properties": {
     "predicate": {
